Move existing tests to legacy namespace
