+++ phpunit.xml.dist
@@ -13,7 +13,7 @@
             <directory suffix="Test.php">tests/integration</directory>
         </testsuite>
         <testsuite name="Legacy Tests">
-            <directory suffix="Test.php">tests</directory>
+            <directory suffix="Test.php">tests/legacy</directory>
         </testsuite>
     </testsuites>
     <filter>
