+++ tests/legacy/UnitTest.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace lewiscowles\Utils\FileSystem\Extension\Tests;
+namespace lewiscowles\Utils\FileSystem\Extension\Tests\Legacy;
 
 use function lewiscowles\Utils\FileSystem\Extension\{extensionIsTooSmall,hasNoExtension,dualPartExtension};
 use PHPUnit\Framework\TestCase;
